backend/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.2
Flask-CORS==3.0.10
redis==4.3.6
PyMySQL==1.0.2
python-dotenv==0.20.0
pillow==8.4.0
numpy==1.19.5
requests==2.27.1
gunicorn==20.1.0
gevent==20.9.0
